committing subscription changes

+++ UNCG-Fitness-Connect/src/main/java/com/UNCG_Fitness/UNCG_Fitness_Connect/subscription/SubscriptionRepository.java
@@ -9,11 +9,11 @@ import java.util.List;
 @Repository
 public interface SubscriptionRepository extends JpaRepository<Subscription, Long> {
     @Query(value = "select * from subscriptions where user_id = :userId", nativeQuery = true)
-    List<Subscription> getSubscriptionsByUser(long userId);
+    List<Subscription> getSubscriptionsByUser(int userId);
 
     @Query(value = "select * from subscriptions where class_id = :classId", nativeQuery = true)
     List<Subscription> getSubscriptionsByClass(int classId);
 
     @Query(value = "select * from subscriptions where class_id = :classId and user_id = :userId", nativeQuery = true)
-    Subscription getOneSubscription(int classId, long userId);
+    Subscription getOneSubscription(int classId, int userId);
 }
